Implementando listagem por user e deletando user

+++ src/controllers/HouseController.js
@@ -7,11 +7,17 @@ class HouseController {
     async index (req, res){
         console.log(req.user._doc)
 
-        // const { status } = req.query
-
         const houses = await House.find()
 
         return res.json({ houses })
+    }
+
+    async shows (req, res) {
+        const { user_id } = req.headers
+
+        const houses = await House.find({user: user_id})
+
+        return res.json({houses})
     }
 
     async store(req, res){
@@ -27,13 +33,10 @@ class HouseController {
             status: status,
         })  
 
-        return res.json({ house })
-    
+        return res.json({ house })    
     }
 
     async update (req, res){
-
-        //Tentando rodar o git bash
 
         console.log('Controller de Alteração Iniciando')
 
@@ -43,16 +46,15 @@ class HouseController {
         const { description,price,location,status } = req.body
 
         console.log('Controller de Alteração: pegando requisições na URL')
-        // console.log('House ID: ' + house_id)
+        console.log('House ID: ' + house_id)
 
+        const house = await House.findById({ house_id })
+        const user = await User.findById({ user_id })
 
-        // const house = await House.findById({ house_id })
-        // const user = await User.findById({ user_id })
-
-        // if (String(user._id) !== String(house.user)){
-        //     console.log('Verificando usuário')
-        //     return res.status(401).json({message: 'User Invalid'})
-        // }
+        if (String(user._id) !== String(house.user)){
+            console.log('Verificando usuário')
+            return res.status(401).json({message: 'User Invalid'})
+        }
 
         const houses = await House.updateOne({ _id: house_id },{
             user: user_id,
@@ -63,8 +65,6 @@ class HouseController {
             status: status,
         })
 
-
-        // return res.send().json({ message: `Casa: ${house.description} alterada com sucesso` })
         return res.json({ houses })
 
     }
@@ -72,6 +72,13 @@ class HouseController {
     async destroy (req, res) {
         const { house_id } = req.body
         const { user_id } = req.headers 
+
+        const house = await House.findById({ _id: house_id})
+
+        if (String(user_id) !== String(house.user)){
+            console.log('Verificando usuário')
+            return res.status(401).json({message: 'User Invalid'})
+        }
 
         await House.findByIdAndDelete({ _id: house_id})
